Split parking.rs (778 lines) into parking/ subdirectory
Split the monolithic parking.rs into smaller, focused files:
- parking/constants.rs (68 lines): constants, ratios, and per-zone cost tables
- parking/state.rs (200 lines): policy state, effects, computation functions, system, saveable, plugin
- parking/tests_constants.rs (180 lines): tests for constants and ratios
- parking/tests_state.rs (317 lines): tests for policy state and computation
- parking/mod.rs (35 lines): module docs and re-exports

All public items are re-exported from mod.rs so external callers are unaffected.

Part of #1389
